pyproject.toml
[project]
name = "fitb-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.62",
    "django-storages>=1.14.4",
    "django>=5.1.3",
    "loguru>=0.7.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
    "django-ninja>=1.3.0",
    "uvicorn>=0.32.1",
]
